roll back refactor into DbCache trait
